requirements.txt
customtkinter
requests
beautifulsoup4
PyPDF2
python-dotenv
faiss-cpu
tqdm
sentence-transformers